requirements.txt
# Optimized requirements for BLIP2 image captioning
# This file is for local development - Docker uses builder/requirements.txt

# Core ML dependencies  
torch>=2.0.0,<2.3.0
torchvision>=0.15.0,<0.18.0
transformers>=4.30.0,<5.0.0

# Image processing
pillow>=9.0.0,<11.0.0

# RunPod serverless
runpod>=1.0.0,<2.0.0

# LAVIS for BLIP2 - using PyPI version for stability
salesforce-lavis>=1.0.0,<2.0.0

# NumPy compatibility fix
numpy<2

# Optional performance dependencies
opencv-python-headless>=4.8.0,<5.0.0
